Added version checks (HALO-1555)

+++ Source/Helpers/ImageURLOptions.swift
@@ -53,7 +53,7 @@ public enum ImageRotation {
     case Ignore
     case VerticalFlip
     case HorizontalFlip
-    case Custom(Int)
+    case Custom(Float)
     
     public var rawValue: String {
         switch self {
